requirements.txt
functions-framework==3.3.0
google-cloud-resource-manager==1.10.4
google-cloud-pubsub==2.18.4
google-auth==2.23.4
google-cloud-logging==3.8.0
google-cloud-datastore==2.18.0